[refactor] create auth util to validate password & check admin
