+++ app/src/main/java/dev/abelab/rippy/service/UserService.java
@@ -15,6 +15,7 @@ import dev.abelab.rippy.api.response.UsersResponse;
 import dev.abelab.rippy.repository.UserRepository;
 import dev.abelab.rippy.logic.UserLogic;
 import dev.abelab.rippy.logic.UserRoleLogic;
+import dev.abelab.rippy.util.AuthUtil;
 
 @RequiredArgsConstructor
 @Service
@@ -38,7 +39,7 @@ public class UserService {
     @Transactional
     public UsersResponse getUsers(final User loginUser) {
         // 管理者かチェック
-        this.userLogic.checkAdmin(loginUser.getId());
+        AuthUtil.checkAdmin(loginUser);
 
         // ユーザ一覧の取得
         final var users = this.userRepository.selectAll();
@@ -59,13 +60,13 @@ public class UserService {
     @Transactional
     public void createUser(final UserCreateRequest requestBody, final User loginUser) {
         // 管理者かチェック
-        this.userLogic.checkAdmin(loginUser.getId());
+        AuthUtil.checkAdmin(loginUser);
 
         // 有効なユーザロールかチェック
         this.usreRoleLogic.checkForValidRoleId(requestBody.getRoleId());
 
         // 有効なパスワードかチェック
-        this.userLogic.validatePassword(requestBody.getPassword());
+        AuthUtil.validatePassword(requestBody.getPassword());
 
         // ユーザを作成
         final var user = this.modelMapper.map(requestBody, User.class);
@@ -85,7 +86,7 @@ public class UserService {
     @Transactional
     public void updateUser(final int userId, final UserUpdateRequest requestBody, final User loginUser) {
         // 管理者かチェック
-        this.userLogic.checkAdmin(loginUser.getId());
+        AuthUtil.checkAdmin(loginUser);
 
         // ユーザを更新
         final var user = this.userRepository.selectById(userId);
@@ -107,7 +108,7 @@ public class UserService {
     @Transactional
     public void deleteUser(final int userId, final User loginUser) {
         // 管理者かチェック
-        this.userLogic.checkAdmin(loginUser.getId());
+        AuthUtil.checkAdmin(loginUser);
 
         // ユーザを削除
         this.userRepository.deleteById(userId);
