+++ app/src/test/java/dev/abelab/rippy/util/AbstractUtil_UT.java
@@ -1,0 +1,14 @@
+package dev.abelab.rippy.util;
+
+import dev.abelab.rippy.annotation.UnitTest;
+
+/**
+ * Abstract Util Unit Test
+ */
+@UnitTest
+public abstract class AbstractUtil_UT {
+
+    static final int SAMPLE_INT = 1;
+	static final String SAMPLE_STR = "SAMPLE";
+
+}
